project-of-java
hello ! this is a java project the name of the project is University Management System.
so in this project I have used java Core concept all work is done on the GUI
so this project started from the the splash screen that is load in that i have used Threading Concept next is open the Admin login frame because whole System is in under administration. as a  software engineering student I believe that Learning with project clear your more concept in this project you have upload the students details ad well as faculty details i have also done more concepts like updated marks leave updated in this project also have notepad and calculater .